feat(database): plugins can now register their own database seeds
Plugins can now use seeds,database hook to register their own seeders.

Fixes #11129
