+++ mod/blog/start.php
@@ -59,6 +59,9 @@ function blog_init() {
 
 	// allow to be liked
 	elgg_register_plugin_hook_handler('likes:is_likable', 'object:blog', 'Elgg\Values::getTrue');
+
+	// register database seed
+	elgg_register_plugin_hook_handler('seeds', 'database', 'blog_register_db_seeds');
 }
 
 /**
@@ -96,35 +99,35 @@ function blog_page_handler($page) {
 	switch ($page_type) {
 		case 'owner':
 			$resource_vars['username'] = elgg_extract(1, $page);
-			
+
 			echo elgg_view_resource('blog/owner', $resource_vars);
 			break;
 		case 'friends':
 			$resource_vars['username'] = elgg_extract(1, $page);
-			
+
 			echo elgg_view_resource('blog/friends', $resource_vars);
 			break;
 		case 'archive':
 			$resource_vars['username'] = elgg_extract(1, $page);
 			$resource_vars['lower'] = elgg_extract(2, $page);
 			$resource_vars['upper'] = elgg_extract(3, $page);
-			
+
 			echo elgg_view_resource('blog/archive', $resource_vars);
 			break;
 		case 'view':
 			$resource_vars['guid'] = elgg_extract(1, $page);
-			
+
 			echo elgg_view_resource('blog/view', $resource_vars);
 			break;
 		case 'add':
 			$resource_vars['guid'] = elgg_extract(1, $page);
-			
+
 			echo elgg_view_resource('blog/add', $resource_vars);
 			break;
 		case 'edit':
 			$resource_vars['guid'] = elgg_extract(1, $page);
 			$resource_vars['revision'] = elgg_extract(2, $page);
-			
+
 			echo elgg_view_resource('blog/edit', $resource_vars);
 			break;
 		case 'group':
@@ -132,7 +135,7 @@ function blog_page_handler($page) {
 			$resource_vars['subpage'] = elgg_extract(2, $page);
 			$resource_vars['lower'] = elgg_extract(3, $page);
 			$resource_vars['upper'] = elgg_extract(4, $page);
-			
+
 			echo elgg_view_resource('blog/group', $resource_vars);
 			break;
 		case 'all':
@@ -159,7 +162,7 @@ function blog_set_url($hook, $type, $url, $params) {
 	if (!elgg_instanceof($entity, 'object', 'blog')) {
 		return;
 	}
-	
+
 	$friendly_title = elgg_get_friendly_title($entity->title);
 	return "blog/view/{$entity->guid}/$friendly_title";
 }
@@ -225,25 +228,25 @@ function blog_archive_menu_setup($hook, $type, $return, $params) {
 	if (empty($page_owner)) {
 		return;
 	}
-	
+
 	$dates = get_entity_dates('object', 'blog', $page_owner->getGUID());
 	if (!$dates) {
 		return;
 	}
-	
+
 	$dates = array_reverse($dates);
-	
+
 	if (elgg_instanceof($page_owner, 'user')) {
 		$url_segment = 'blog/archive/' . $page_owner->username;
 	} else {
 		$url_segment = 'blog/group/' . $page_owner->getGUID() . '/archive';
 	}
-	
+
 	$years = [];
 	foreach ($dates as $date) {
 		$timestamplow = mktime(0, 0, 0, substr($date, 4, 2), 1, substr($date, 0, 4));
 		$timestamphigh = mktime(0, 0, 0, ((int) substr($date, 4, 2)) + 1, 1, substr($date, 0, 4));
-	
+
 		$year = substr($date, 0, 4);
 		if (!in_array($year, $years)) {
 			$return[] = ElggMenuItem::factory([
@@ -255,10 +258,10 @@ function blog_archive_menu_setup($hook, $type, $return, $params) {
 				]
 			]);
 		}
-		
+
 		$link = $url_segment . '/' . $timestamplow . '/' . $timestamphigh;
 		$month = trim(elgg_echo('date:month:' . substr($date, 4, 2), ['']));
-		
+
 		$return[] = ElggMenuItem::factory([
 			'name' => $date,
 			'text' => $month,
@@ -295,7 +298,7 @@ function blog_prepare_notification($hook, $type, $notification, $params) {
 	], $language);
 	$notification->summary = elgg_echo('blog:notify:summary', [$entity->title], $language);
 	$notification->url = $entity->getURL();
-	
+
 	return $notification;
 }
 
@@ -306,4 +309,21 @@ function blog_ecml_views_hook($hook, $entity_type, $return_value, $params) {
 	$return_value['object/blog'] = elgg_echo('blog:blogs');
 
 	return $return_value;
+}
+
+/**
+ * Register database seed
+ *
+ * @elgg_plugin_hook seeds database
+ *
+ * @param \Elgg\Hook $hook Hook
+ * @return array
+ */
+function blog_register_db_seeds(\Elgg\Hook $hook) {
+
+	$seeds = $hook->getValue();
+
+	$seeds[] = \Elgg\Blog\Seeder::class;
+
+	return $seeds;
 }
